// vue-app/src/types/global.ts
export interface Attachment {
  name: string
  content: string
  startWordIndex: number
  endWordIndex: number
}